+++ lib/__init__.py
@@ -1,0 +1,5 @@
+from armor import Armor
+from character import Character
+from enemy import Enemy
+from weapon import Weapon
+from game import Game
