+++ main.py
@@ -1,0 +1,5 @@
+from lib import Game
+
+if __name__ == '__main__':
+    game = Game()
+    game.start()
